서버에서 몬스터가 정상적으로 삭제되지 않는 문제 수정
정상적으로 dead 애니메이션 실행후 클라이언트에서는 사라지지만 충돌범위는 그대로 남아있는 현상을 수정하였다.

함수 객체가 캡쳐하는 shared_ptr 오브젝트가 문제였으며 해당 부분을 weak_ptr로 수정하여 해결하였다

+++ SandyHeroesServer/Scene.cpp
@@ -14,23 +14,35 @@ XMVECTOR Scene::GetPickingPointAtWorld(float sx, float sy, Object* picked_object
 
 void Scene::AddObject(std::shared_ptr<Object> object)
 {
+	if(is_updating_objects_)
+	{
+		add_object_list_.push_back(object);
+		return;
+	}
 	object_list_.push_back(object);
 }
 
 void Scene::DeleteObject(std::shared_ptr<Object> object)
 {
+
+	if(is_updating_objects_)
+	{
+		delete_object_list_.push_back(object);
+		return;
+	}
 	for (auto& sector : sectors_)
 	{
 		sector.DeleteObject(object.get());
 	}
 
 	object_list_.remove(object);
+	std::cout << "Object " << object->name() << " with id " << object->id() << " is destroyed by Scene\n";
+
 }
 
 void Scene::Update(float elapsed_time)
 {
 	is_updating_objects_ = true;
-
 	for (const std::shared_ptr<Object>& object : object_list_)
 	{
 		object->Update(elapsed_time);
@@ -46,6 +58,7 @@ void Scene::Update(float elapsed_time)
 	for (const std::shared_ptr<Object>& object : delete_object_list_)
 	{
 		object_list_.remove(object);
+		std::cout << "Object " << object->name() << " with id " << object->id() << " is destroyed by Scene\n";
 	}
 	delete_object_list_.clear();
